Enable caching last-token-updated timestamp
Summary:
There was only a single method for setting the access token on the
Facebook class (setAccessToken()).  This did not distinguish between
setting a token loaded from cache vs.  setting a new token resulting
from authorization.

Since setAccessToken() updates the last-token-updated timestamp, and
since this timestamp is used to determine when to refresh the token,
it always looked like the token had just been updated when the app's
Activity started.  This meant the Facebook class would never try to
refresh the token.

To fix this, the Facebook class now supports getting the value of the
last-token-updated timestamp, and also supports restoring this along
with the token value and expiration timestamp in a single operation.

This also updates Hackbook to demonstrate caching.

Test Plan:
- ant build Hackbook
- Ran Hackbook, verified that we could get and restore the timestamp.

Revert Plan:
safe to revert

Tags:

Reviewers: clang, vijaye, caabernathy

Reviewed By: clang

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D552473

Task ID: 1363221